Fix point being moved by window and tab-bar
We insert lines, which messes with (point), and both the windows and tab-bar
keep their own record of (point) and restore it when switching to the buffer.

As a result we need to keep our own marker to where point should be, either
through our insertion or user commands (moving point), and then *we* restore it
when the buffer is redisplayed.